requirements.txt
streamlit>=1.40.0
chromadb>=0.5.0
transformers>=4.40.0
torch
Pillow>=10.0
google-genai>=1.0.0
pydantic-settings>=2.0
